auth renamed to authorizationService
